Update user's weight
Creates a weightSubmit function used on the weighIn page. Searches
through the users array of objects for a name match. If there is a name
match, the weight parameter is subtracted from the startWeight to create
a new weightLoss for the matched array element. The getWeightLoss global
method is called to transform updated users array into the myArray
variable (everything without the players weight and sent to the
controller to update the Player Ranking Page).