#14 Feature: Random Inspirational Quote, bug fixs
Feature:
- This data will be feature at the bottom of the Weigh-In page under the
re-titled sub-title "Random Inspirational Quote of the Day". It should
be randomly selected and downloaded to client.
- The quotes will be placed in an $scope.inspiration array. A function
called getRandomQuote is use to randomly select an index in the
inspiration array and assign the selected quote to the varible
randomQoute. The randomQuote varible is placed in a paragraph element
under the sub-title.
Bugs:
- on css page, word "random" in the comments is mispelled
- on Ranking page, change Weight loss title to Total Weight Loss
- change color of upOrDown varible in the css but seen on Ranking page