#17 Display Current Week Number and fixed bugs
Feature purpose is to display to the currentUser the current week number
of the challenge.
- This number is displayed on the Ranking page below the title and on
the weighIn page under the navigation.
- A global startDate and currentWeek variable is set when the app is
initalize. When the currentUser logs in, a currentDate variable is made
of today's date.  A function called getCurrentWeekNumber() subtracts
currentDate from startDate to get the numberOfDays variable. This
varible is divided by 7 to get the currentWeek variable which is display
on the html page.
- The color of the current week number is black instead of orange to
keep the focus on the weighIn area and player rankings. It is also
center align similar to the page title.

Bugs:
- add comment to getCurrentWeightLoss()  called in weighIn function
- add "lbs" to upOrDown variables on html page
- add "Page" to the Player Ranking Title on the Ranking Page