Toggle Switch for Weight Loss or Percentage Loss
Issue #28 : Creates a check-box on the Ranking Page to allow the user to
see weight loss either by pounds or percentage loss.
-  While newArray is being built, each userData array element weightLoss
is divided by the startWeight to create a new attribute,
currentPercentLoss.
- a global variable name percentOrPounds is ititialize to false at
startup
- On the Ranking Page's total weight loss section, if the checkbox is
not check then the ng-show span element containing the weightLoss is
displayed. If the checkbox is checked, then the the ng-show span element
containing currentPercentLoss is displayed.
- The percentage is displayed using a number filter, faction set to 1,
and multiplied by 100.