#33 Fixed several Nav. issues; Add new player bugs
1.  In mobile view, the "Join Challenge" nav link was leaking out of the
nav box into a white area.
- Fixed by moving the "Join Challenge" nav link to a button link on the
Log-In Page.
- Return navigation to original two 6 column nav links.
2. Access control for the site's Ranking and Weigh-In page
- implemented global variable enableNav that is set to false initially.
When this is false, the setView method is disable, effectively not
allowing switching of views. It is set to true only if log-In is
successful.
3. Fixed alot of bugs
- when "Join Challenge" button is press the page would not be enable, or
it would show both logIn and setUpPage pages. Solved by creating a
seperate "join" view function and unableing this view once a new player
has been created. The setView function was tweak to un-show the
setUpPage page when the logIn page is shown. Also upon login the
enableNav variable changed to true.
- When a new player is created the current week would not display.
Solved by making the getCurrentWeekNumber function global.
- After a new player is made and switch to the weighIn page, the
submitWeight button would not switch from weighIn to Ranking page. The
information would be update, just the player is stuck on the weighIn
page. Solved by making the  getCurrentWeightLoss function global.