Continue Guerrilla Testing Part 2
The following changes was made from reccomendations:
1. Log-out Feature:
Solution: Inside a CSS coded sticky footer, use a red color sign-out
picture with a ng-click logOut () function attached to it. This function
clears manu global variables including the currentUser and newArray
(holds the client side user data). The function will lock down the
navigation with the enableNav variable and redirect to the logInPage.