Winner Function and one bug
This function will determine if the user has passed the 10 pound
threshold. If so, then will update the the userData array to let eveyone
know the winner.
Solution:
1. Add to the userData array model the "winner" boolean variable as
false.
2. Add to the client side newArray array "winner" the  boolean variable
as false.
3. A span html element is made with a ng-show directive looking for a
winner variable. If the winner attribute of a user on the Ranking page
is true, then the span is enable.
4. A updateWinner function was coded to update the userData array model.
The currentUser is matched to an userData array and the winner attribute
is change to true.
5. Inside the weighIn function, an if statment checks if the currentUser
weightLoss is equal or more then 10. If true, then the updateWinner
Function is run. The getWeightLoss function is run next to create the
client side user array data. The Ranking Page is updated with the Winner
shown.
6. Bug: the newArray should update in the log in function.