#36 Break main Controller into smaller controllers
Continue objective to break entire app into smaller pieces.
Solution:
1. identify functions and variables that are shared thoughout app or
retrieves array based data. Place this information into services.Ensure
each service has one function or goal.
- Time service: Returns shared resources: today's date, minutes in a
day, and start of challenge date.
- SetView service: Controls the authoration of the navigation links and
moving between pages.  Returns shared resources: enableNav and join. The
setView method returns a page name to the scope.view varible linked to
the ng-show on each html page.
- CurrentWeek service use the Time service: returns the current week of
the challenge. This is used on the weigh-in and ranking page. The start
date of the challenge is subtracted from today's date and divided by
minutes in a day. This amount is rounded to a whole number and used in
the currentWeek variable.
- NewUser service: Returns and updates the shared resource, current user
name.
- Cred service: Returns and updates the shared resource, the log-in cred
array.
- UserData service use the Time service: Returns  and updates the shared
resource userData array to other services. Have the addUser method to
add indiviual users.
- UpdateUserData service uses the UserData, NewUser, Time, and
CurrentWeightLoss service: This service is used on the weigh-in page. It
takes the inputed weight of the user to update the weight loss, change
the lastUpdate status, and check if the user has won. The
CurrentWeightLoss service is use to determine if neg. or positive
weightloss.
- RandomQuote service: returns a random quote from the inspiration
array.
- NewArray service use the UserData, Time, and NewUser services. This
service returns a shared resource, newArray, that is displayed for the
current user. This is the userData array strip of the start weight but
adds the percentLoss.
- CurrentWeightLoss service uses the NewArray and NewUser services. This
service gets, updates, and returns the upOrDown shared resource that
links to the ng-show (how much loss) on the ranking page.
Extra:
New scope method navigation: Use the Setview service to update the
scope.view variable and change the pages.
New scope method getRandomQuote: Use the RandomQuote service to update
scope.randomQuote variable.